Ver.2.19.95
System
- Position.java
 Positionの座標がfloatで管理されるようになった。

New Property
- Rotate.java
 指定した角度までアニメーションオブジェクトを瞬時に回転させる

+++ src/main/java/org/ayato/animation/MoveTo.java
@@ -6,7 +6,7 @@ import java.awt.*;
 
 public class MoveTo extends TimeConverter{
     private final int x, y;
-    private int dx, dy, currentX, currentY;
+    private float dx, dy, currentX, currentY;
     private final VelocityFormat format;
     public MoveTo(int x, int y, long maxTime, VelocityFormat format1) {
         super(maxTime);
